==> Step1_LabelReadmissions.ipynb <==
# Prediction of Hospital Readmissions

This project's goal is to reproduce the claims from Zebin and Chaussalet's paper, 'Design and implementation of a deep recurrent model for prediction of readadmission in urgent care using electronic health records'.

These first series of notebooks are for processing, cleaning, and sorting data.


## Claims

When predicting ICU readmissions:

1. LSTM+CNN produced higher accuracy than logistic regression, random forest, and SVM.
2. LSTM+CNN produced higher precision than logistic regression, random forest, and SVM.
3. LSTM+CNN produced higher recall than logistic regression and SVM

# Notebook 1

This notebook gathers the unique ICUSTAY_IDs, then filters out the cases for valid readmissions.

This is accomplished by combining the data from ADMISSION, ICUSTAYS, and PATIENTS tables from the MIMIC-III database.

import numpy as np
import pandas as pd

patients = pd.read_csv('./mimic-iii/PATIENTS.csv')
patients.head(1)

admissions = pd.read_csv('./mimic-iii/ADMISSIONS.csv')
print(admissions.shape)
admissions.head(1)

transfers = pd.read_csv('./mimic-iii/TRANSFERS.csv')
transfers.head(1)

icustays = pd.read_csv('./mimic-iii/ICUSTAYS.csv')
icustays.head(1)


# print(len(icustays.groupby(['HADM_ID'])))
# test = (icustays.groupby(['HADM_ID']).size() > 1).reset_index()
# test[test[0]==True]

## NOTE: Organ Donors

There are some cases where a patient as multiple inconsistent death times in duplicated records. Either the patient has 2 different death times or the admit and discharge times do not align with death time.
It appears that the organ donor collection is processed under a different HADM_ID number and the entries are inconsistent.

# addmissions where DEATH IN HOSPITAL
organ = admissions[admissions['HOSPITAL_EXPIRE_FLAG']==1].reset_index(drop=True)

organ[organ.duplicated('SUBJECT_ID',keep=False)].sort_values('SUBJECT_ID').head(2)

## Example

Patient with `SUBJECT_ID = 250`

# patients[patients['SUBJECT_ID'] == 250]

# admissions[admissions['SUBJECT_ID'] == 250]

# transfers[transfers['SUBJECT_ID'] == 250]

## Create Dataset

SUBJECT_ID exmaples  291, 283, 250

patients = patients.drop('ROW_ID', axis=1)
admissions = admissions.drop('ROW_ID', axis=1)
icustays = icustays.drop('ROW_ID', axis=1)

patients_columns = ['SUBJECT_ID', 'GENDER', 'DOB', 'DOD', 'EXPIRE_FLAG']
admissions_columns = ['SUBJECT_ID', 'HADM_ID', 'ADMITTIME', 'DISCHTIME', 'DEATHTIME',
                    'INSURANCE', 'LANGUAGE', 'RELIGION', 'MARITAL_STATUS', 'ETHNICITY',
                    'DIAGNOSIS', 'HOSPITAL_EXPIRE_FLAG', 'HAS_CHARTEVENTS_DATA']
icustays_columns = ['SUBJECT_ID','HADM_ID','ICUSTAY_ID','FIRST_CAREUNIT',
                    'LAST_CAREUNIT','FIRST_WARDID','LAST_WARDID','INTIME','OUTTIME','LOS']

dataset = pd.merge(icustays[icustays_columns],admissions[admissions_columns],'left',['SUBJECT_ID','HADM_ID'])
dataset = dataset.merge(patients[patients_columns],'left',['SUBJECT_ID'])

dataset.columns

print(dataset.shape)
dataset.head(1)

dataset['DOB'] = pd.to_datetime(dataset['DOB'], errors='coerce')
dataset['DOD'] = pd.to_datetime(dataset['DOD'], errors='coerce')
dataset['ADMITTIME'] = pd.to_datetime(dataset['ADMITTIME'], errors='coerce')
dataset['DISCHTIME'] = pd.to_datetime(dataset['DISCHTIME'], errors='coerce')
dataset['DEATHTIME'] = pd.to_datetime(dataset['DEATHTIME'], errors='coerce')
dataset['INTIME'] = pd.to_datetime(dataset['INTIME'], errors='coerce')
dataset['OUTTIME'] = pd.to_datetime(dataset['OUTTIME'], errors='coerce')


dataset_columns = ['SUBJECT_ID', 'HADM_ID', 'ICUSTAY_ID',  'ADMITTIME', 'DISCHTIME', 'DEATHTIME', 
                   'INTIME', 'OUTTIME','DOB', 'DOD', 'EXPIRE_FLAG', 'HOSPITAL_EXPIRE_FLAG', 
                   'HAS_CHARTEVENTS_DATA', 'FIRST_CAREUNIT', 'LAST_CAREUNIT', 
                   'FIRST_WARDID', 'LAST_WARDID', 'LOS', 'INSURANCE', 'LANGUAGE',
                   'RELIGION', 'MARITAL_STATUS', 'ETHNICITY', 'DIAGNOSIS', 'GENDER']

dataset = dataset[dataset_columns]

dataset = dataset[(dataset['DIAGNOSIS'] != 'ORGAN DONOR ACCOUNT') & (dataset['DIAGNOSIS'] != 'DONOR ACCOUNT') & 
                  (dataset['DIAGNOSIS'] != 'ORGAN DONOR') & (dataset['DIAGNOSIS'].notnull())].reset_index(drop=True)

## Goal: Find cases of readmission

These are the numbers from the journal paper that this project is reproducing.

1. Transferred & returned               = 3555
2. Transferred & died                   = 1974
3. Discharged & returned < 30 days      = 3205
4. Discharged & died < 30 days          = 2556


### 1. Transferred & Returned

```python
dataset['TRANSFERRED_RETURNED'] = icustays.duplicated(subset=['HADM_ID']).astype(int)
```

dataset = dataset.sort_values(['HADM_ID','ADMITTIME','INTIME']).reset_index(drop=True)
# same HADM_ID, but diff ICUSTAY_ID
dataset['ICU_VISIT_PER_ADMIT'] = dataset.groupby('HADM_ID')['ICUSTAY_ID'].cumcount()
dataset['TRANSFERRED_RETURNED'] = 0
dataset.loc[dataset['ICU_VISIT_PER_ADMIT'] > 0, 'TRANSFERRED_RETURNED'] = 1

np.sum(dataset['TRANSFERRED_RETURNED']).astype(int)

### 2. Transferred & Died

```python
dataset['TRANSFERRED_DEATH'] = admissions['HOSPITAL_EXPIRE_FLAG'] and not dataset['DEATH_IN_ICU']
```

### Define death

```python
patients['EXPIRE_FLAG'] == 1
```

### Define death in hospital

```python
admissions['HOSPITAL_EXPIRE_FLAG'] == 1
```

### Define death in ICU (merge addmissions & icustays)

```python
dataset['DEATH_IN_ICU'] = (admissions['DEATHTIME'] > icustays['INTIME']) & (admissions['DEATHTIME'] < icustays['OUTTIME']).astype(int)
```

dataset['TRANSFERRED_DEATH'] = ((dataset['DEATHTIME'].notnull()) & 
                                (dataset['HOSPITAL_EXPIRE_FLAG'] == 1) & 
                                ((dataset['DEATHTIME'] < dataset['INTIME']) | 
                                 (dataset['DEATHTIME'] > dataset['OUTTIME']))).astype(int)

np.sum(dataset['TRANSFERRED_DEATH']).astype(int)

### 3. Discharged & Returned < 30

This is the wrong way. This labels the return visit instead of discharged visit

```python
sort admissions by SUBJECT_ID, ADMITTIME
shifted = admissions.shift(1)
shifted['DISCHTIME'] = pd.to_datetime(shifted['DISCHTIME'], errors='coerce')
dataset['DISCHARGED_RETURNED'] = (shifted['SUBJECT_ID'] == admissions['SUBJECT_ID']) & (admissions['ADMITTIME'] - shifted['DISCHTIME'] <= np.timedelta64(30, 'D')) 
```

dataset.sort_values(by=['SUBJECT_ID', 'DISCHTIME'], ignore_index=True, inplace=True)
shifted = dataset.shift(-1)
dataset['DISCHARGED_RETURNED'] = ((shifted['SUBJECT_ID'] == dataset['SUBJECT_ID']) & 
                                  (shifted['HADM_ID'] != dataset['HADM_ID']) & 
                                  (shifted['INTIME'] - dataset['DISCHTIME'] < np.timedelta64(30, 'D'))).astype(int)

np.sum(dataset['DISCHARGED_RETURNED']).astype(int)

### 4. Discharged & Died < 30 (merge patients, addmisions)

Need last discharged visit HADM_ID

```python
dataset['DISCHARGED_DEATH'] = patients['EXPIRE_FLAG'] & not admissions['HOSPITAL_EXPIRE_FLAG'] & (patients['DOD'] - admissions['DISCHTIME'] <= np.timedelta64(30, 'D')) 
```

# dataset.sort_values(by=['SUBJECT_ID', 'DISCHTIME'], ignore_index=True, inplace=True)
dataset['DISCHARGED_DEATH'] = ((dataset['EXPIRE_FLAG']==1) & (dataset['HOSPITAL_EXPIRE_FLAG']==0) & 
                               (dataset['DOD'].notnull()) & 
                               (dataset['DOD'] - dataset['DISCHTIME'] <= np.timedelta64(30, 'D')) ).astype(int)

np.sum(dataset['DISCHARGED_DEATH']).astype(int)

## Age

dataset['AGE'] = ((pd.to_datetime(dataset['ADMITTIME']).dt.date - pd.to_datetime(dataset['DOB']).dt.date) 
                  / np.timedelta64(1, 'Y')).astype(int)

# dataset = dataset[(dataset['TRANSFERRED_RETURNED'] == 1) | 
#                   (dataset['TRANSFERRED_DEATH'] == 1) | 
#                   (dataset['DISCHARGED_RETURNED'] == 1) | 
#                   (dataset['DISCHARGED_DEATH'] == 1) |
#                   (dataset['AGE'] >= 18)]
dataset = dataset[(dataset['AGE'] >= 18)]

## Remove Died in ICU

print(dataset.shape)
dataset['DIED_IN_ICU'] = ((dataset['DEATHTIME'].notnull()) & 
                                (dataset['HOSPITAL_EXPIRE_FLAG'] == 1) & 
                                ((dataset['DEATHTIME'] >= dataset['INTIME']) & 
                                 (dataset['DEATHTIME'] <= dataset['OUTTIME']))).astype(int)

dataset = dataset[(dataset['DIED_IN_ICU']==0)]
print(dataset.shape)

## Goal: Find cases of readmission

1. Transferred & returned               = 3555
2. Transferred & died                   = 1974
3. Discharged & returned < 30 days      = 3205
4. Discharged & died < 30 days          = 2556

print(np.sum(dataset['TRANSFERRED_RETURNED']==1))
print(np.sum(dataset['TRANSFERRED_DEATH']==1))
print(np.sum(dataset['DISCHARGED_RETURNED']==1))
print(np.sum(dataset['DISCHARGED_DEATH']==1))

dataset['READMISSION'] = ((dataset['TRANSFERRED_RETURNED']==1) | (dataset['TRANSFERRED_DEATH']==1) | (dataset['DISCHARGED_RETURNED']==1) | (dataset['DISCHARGED_DEATH']==1)).astype(int)

dataset.to_csv('data/readmission.csv')
