==> Step2A_StoreCEByICUSTAY_dask.ipynb <==
# Split Chart Events by ICUSTAY ID

import pandas as pd
import numpy as np
from tqdm.notebook import trange, tqdm
import traceback
import os
import dask.dataframe as dd
from dask.distributed import Client

dirs = os.listdir('data')

if 'parquet' not in dirs:
    os.mkdir('data/parquet')

if 'samples' not in dirs:
    os.mkdir('data/samples')

if 'samples_demographics' not in dirs:
    os.mkdir('data/samples_demographics')

if 'samples_icd9' not in dirs:
    os.mkdir('data/samples_icd9')

if 'samples_numerical' not in dirs:
    os.mkdir('data/samples_numerical')

if 'samples_statB' not in dirs:
    os.mkdir('data/samples_statB')

CHARTEVENTS_FILENAME = 'mimic-iii/CHARTEVENTS.csv'
READMISSION_FILENAME = 'data/readmission.csv'
SAMPLES_DIR = 'data/parquet/'
DATASTORE_FILENAME = 'samples.h5'
FEATHER_EXT = '.feather'
ROWS_TO_READ = 1000000
MAX_ROWS_CHARTEVENTS = 330712483

skip_rows = 0

chartevents_columns = ["ROW_ID","SUBJECT_ID","HADM_ID","ICUSTAY_ID","ITEMID","CHARTTIME","STORETIME","CGID","VALUE","VALUENUM","VALUEUOM","WARNING","ERROR","RESULTSTATUS","STOPPED"]


chartevents_dtype = {'ROW_ID':str,
'SUBJECT_ID':np.float64,
'HADM_ID':np.float64,
'ICUSTAY_ID':np.float64,
'ITEMID':np.float64,
'CHARTTIME':str,
'STORETIME':str,
'CGID':str,
'VALUE':str,
'VALUENUM':str,
'VALUEUOM':str,
'WARNING':str,
'ERROR':str,
'RESULTSTATUS':str,
'STOPPED':str}

readmission = pd.read_csv(READMISSION_FILENAME).sort_values(['SUBJECT_ID','HADM_ID','ICUSTAY_ID']).reset_index(drop=True)


icustay_list = readmission['ICUSTAY_ID'].unique().tolist()
icustay_list.sort()
division_index = np.array(range(icustay_list[0],icustay_list[-1],int(len(icustay_list)/300)))
division_index = np.append(division_index,icustay_list[-1]+1)
division_index = division_index.astype(str)
division_index = np.char.add(division_index,np.array(['_0']*len(division_index)).astype(str))
len(division_index)

client = Client()
client

df = dd.read_csv(CHARTEVENTS_FILENAME, dtype=chartevents_dtype ,assume_missing=True)

df = df.dropna(subset=['SUBJECT_ID','HADM_ID','ICUSTAY_ID','CHARTTIME']).drop('ROW_ID',axis=1)

df['SUBJECT_ID'] = df.map_partitions(lambda x: x['SUBJECT_ID'].astype(np.int64))
df['HADM_ID'] = df.map_partitions(lambda x: x['HADM_ID'].astype(np.int64))
df['ICUSTAY_ID'] = df.map_partitions(lambda x: x['ICUSTAY_ID'].astype(np.int64))
df['CHARTTIME'] = df.map_partitions(lambda x: pd.to_datetime(x['CHARTTIME'], errors='coerce'))# , format='%Y-%m-%d %H:%M:%S')

df.info

valid_chartevents = df.map_partitions(lambda df: dd.multi.merge(readmission[['SUBJECT_ID','HADM_ID','ICUSTAY_ID']],df,'inner',['SUBJECT_ID','HADM_ID','ICUSTAY_ID']))

# This worked
valid_chartevents.to_parquet(SAMPLES_DIR, engine="pyarrow", partition_on='ICUSTAY_ID', compute=True)
